Remove parametrized tests
