+++ src/qibo/tests/test_core_gates.py
@@ -11,7 +11,7 @@ try:
     BACKENDS = ["custom", "defaulteinsum", "matmuleinsum",
                 "numpy_defaulteinsum", "numpy_matmuleinsum"]
 except ModuleNotFoundError: # pragma: no cover
-    BACKENDS = ["defaulteinsum", "matmuleinsum"]
+    BACKENDS = ["numpy_defaulteinsum", "numpy_matmuleinsum"]
 
 
 def apply_gates(gatelist, nqubits=None, initial_state=None):
@@ -71,5 +71,32 @@ def test_z(backend):
     target_state = np.ones_like(final_state) / 2.0
     target_state[2] *= -1.0
     target_state[3] *= -1.0
+    np.testing.assert_allclose(final_state, target_state)
+    qibo.set_backend(original_backend)
+
+
+@pytest.mark.parametrize("backend", BACKENDS)
+def test_rx_parameter_setter(backend):
+    """Check that the parameter setter of RX gate is working properly."""
+    def exact_state(theta):
+        phase = np.exp(1j * theta / 2.0)
+        gate = np.array([[phase.real, -1j * phase.imag],
+                        [-1j * phase.imag, phase.real]])
+        return gate.dot(np.ones(2)) / np.sqrt(2)
+
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    theta = 0.1234
+    gate = gates.RX(0, theta=theta)
+    initial_state = K.cast(np.ones(2) / np.sqrt(2))
+    final_state = gate(initial_state)
+    target_state = exact_state(theta)
+    np.testing.assert_allclose(final_state, target_state)
+
+    theta = 0.4321
+    gate.parameters = theta
+    initial_state = K.cast(np.ones(2) / np.sqrt(2))
+    final_state = gate(initial_state)
+    target_state = exact_state(theta)
     np.testing.assert_allclose(final_state, target_state)
     qibo.set_backend(original_backend)
